Rename and move useWordMatchQuiz hook
