+++ frontend/src/react/Components/WordMatchQuiz.jsx
@@ -2,7 +2,7 @@ import React, { useMemo, useEffect, useState } from 'react'
 import PropTypes from 'prop-types'
 
 import shuffle from '../utils/shuffle'
-import useWordMatchQuiz from '../stores/useWordMatchQuizStore'
+import useWordMatchQuiz from '../../hooks/useWordMatchQuiz'
 import QuizButton from './QuizButton'
 
 const containerStyle = {
